requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.3
astor==0.8.1
asttokens==2.0.8
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.9.0
boto3==1.24.75
botocore==1.27.75
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
debugpy==1.8.11
decorator==5.1.1
dominate==2.7.0
duckdb==1.1.3
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.0.0
fastapi==0.115.6
Flask==2.2.5
Flask-Bootstrap==3.3.7.1
Flask-S3==0.3.3
Frozen-Flask==0.18
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
grpclib==0.4.7
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0
html5==0.0.9
html5lib==1.1
hyperframe==6.0.1
idna==3.10
importlib-metadata==4.12.0
ipykernel==6.29.5
ipython==8.10.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
modal==0.70.1
multidict==6.1.0
mypy-extensions==0.4.3
narwhals==1.19.1
nba_api==1.6.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==2.5.2
plotly==5.24.1
prompt-toolkit==3.0.31
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.15.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.6.0
scikit-learn==1.6.0
scipy==1.13.1
semver==3.0.2
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.5.0
starlette==0.41.3
streamlit==1.12.0
synchronicity==0.9.8
tabulate==0.8.9
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
tornado==6.4.2
traitlets==5.4.0
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.18
validators==0.34.0
visitor==0.1.3
watchfiles==1.0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.8
yarl==1.18.3
zipp==3.8.1
zope.interface==7.2
